+++ terraform/modules/gke/main.tf
@@ -41,8 +41,9 @@ module "gke" {
   ip_range_pods     = var.ip_range_pods_name
   ip_range_services = var.ip_range_services_name
 
-  deletion_protection = var.cluster_deletion_protection
-  service_account     = var.cluster_service_account
+  deletion_protection    = var.cluster_deletion_protection
+  service_account        = var.cluster_service_account
+  create_service_account = var.create_cluster_service_account
 
   node_pools = [
     {
@@ -185,9 +186,10 @@ module "postgres" {
     transaction_log_retention_days = null
   }
 
-  db_name     = var.database_name
-  user_name   = var.database_username
-  user_labels = var.labels
+  db_name              = var.database_name
+  user_name            = var.database_username
+  user_deletion_policy = var.database_user_deletion_policy
+  user_labels          = var.labels
 
   depends_on = [google_service_networking_connection.default]
 }